HACBS-228: Add support for bootstrapping rekor
This commit moves rekor into the bootstrap script. By default the rekor
service is not externally accessible outside the cluster.

When ran with the `preview` flag, the rekor server is configured with a
domain of the cluster in the form of `rekor.<domain>` and the
tekton-chains configmap is patched to use this value as the
`transparency.url` value.

After this PR is merged, executing the `boostrap-cluster.sh` script
again after having ran it with the `preview` flag will return the rekor
instance to not being externally accessible outside of the cluster.

Signed-off-by: Rob Nester <rnester@redhat.com>